Squashed commit of the following:
commit d913fa16a08227b0ac21d8dbc0470a4a534a81f6
Merge: 1a3c2a7 e1bf075
Author: SheeeChaos <146656845+SheeeChaos@users.noreply.github.com>
Date:   Mon Dec 11 09:00:48 2023 +0800

    Merge pull request #12 from SheeeChaos/Group-1-update

    update

commit e1bf0750f9395073cbb4566e5721e25edb083b13
Author: SheeeChaos <146656845+SheeeChaos@users.noreply.github.com>
Date:   Mon Dec 11 08:59:30 2023 +0800

    update

commit 1a3c2a7fdccd469812d8e6ded93ef228d3753b63
Merge: 3371538 a8cafe9
Author: SheeeChaos <146656845+SheeeChaos@users.noreply.github.com>
Date:   Thu Dec 7 23:13:23 2023 +0800

    Merge pull request #10 from SheeeChaos/Group-1-update

    11

commit a8cafe9a49f61f42c3c3c4d56736f582bb1d1472
Merge: f383916 3371538
Author: SheeeChaos <146656845+SheeeChaos@users.noreply.github.com>
Date:   Thu Dec 7 23:13:07 2023 +0800

    Merge branch 'master' into Group-1-update

commit f3839162140a8bbe6dbf29bd752b30cf616910a9
Author: SheeeChaos <146656845+SheeeChaos@users.noreply.github.com>
Date:   Thu Dec 7 23:09:05 2023 +0800

    11

commit 3371538d38095e0e0192348c7cbb57193f378ad7
Merge: abecb1e 45e1ef8
Author: SheeeChaos <146656845+SheeeChaos@users.noreply.github.com>
Date:   Thu Dec 7 23:02:16 2023 +0800

    Merge pull request #9 from MCAces/master

    UI update

commit 45e1ef82fade027d2f7087e46721648e2c0f273d
Author: MCAces <148616331+MCAces@users.noreply.github.com>
Date:   Thu Dec 7 22:21:16 2023 +0800

    UI update

commit abecb1e18933ef49b3809931065331a6b0fb51b4
Merge: eab0213 b833fb2
Author: SheeeChaos <146656845+SheeeChaos@users.noreply.github.com>
Date:   Thu Dec 7 02:39:51 2023 +0800

    Merge pull request #8 from MCAces/master

    Group 4 - update

commit eab0213c26599778792bad8c0eb1fa63405cb8ad
Author: SheeeChaos <146656845+SheeeChaos@users.noreply.github.com>
Date:   Thu Dec 7 02:36:51 2023 +0800

    Added

commit b833fb25e3e40e4f2fa61ec62d8296966d27f84a
Author: MCAces <148616331+MCAces@users.noreply.github.com>
Date:   Thu Dec 7 02:24:28 2023 +0800

    Group 4 - update

commit 18f243e17ffd6060b07c29b5fb98b813695af4bd
Author: SheeeChaos <146656845+SheeeChaos@users.noreply.github.com>
Date:   Tue Dec 5 23:44:28 2023 +0800

    Group 1 update 1.0

commit b8e6dbc0ad60e99a38f72cf1186dfe2feb2cffbd
Merge: ffa6950 7c6035f
Author: SheeeChaos <146656845+SheeeChaos@users.noreply.github.com>
Date:   Mon Dec 4 17:09:08 2023 +0800

    Merge pull request #7 from MCAces/master

    testing2

commit 7c6035f68009d8bde65ea2cc8e37dee00af8e582
Author: MCAces <148616331+MCAces@users.noreply.github.com>
Date:   Mon Dec 4 16:40:21 2023 +0800

    testing

commit ffa6950e8d5bdf426e7d4316053ffe51025a0c9c
Merge: 39a474c 8039ca0
Author: SheeeChaos <146656845+SheeeChaos@users.noreply.github.com>
Date:   Mon Dec 4 16:33:34 2023 +0800

    Merge pull request #6 from MCAces/master

    testing

commit 8039ca0653c428e80cbc490772ebeec63e2b7abc
Author: MCAces <148616331+MCAces@users.noreply.github.com>
Date:   Mon Dec 4 16:29:01 2023 +0800

    testing

+++ Models/Product_Info.cs
@@ -32,7 +32,7 @@ namespace SBIT3J_SuperSystem_Final.Models
         public string Color { get; set; }
         public Nullable<bool> Sex { get; set; }
         public Nullable<bool> Is_Archived { get; set; }
-        public Nullable<int> Price { get; set; }
+        public Nullable<decimal> Price { get; set; }
         public Nullable<int> Stock_Level { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
